edge-ai-void-detection/modules/processimages/requirements.txt
azure-iothub-device-client==1.4.0
tensorflow
tensorflow-serving-api
grpcio
statistics
matplotlib
azure-storage
azure-storage-blob
